backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pdfplumber==0.11.4
python-docx==1.1.2
transformers==4.41.2
torch==2.3.1
sentencepiece==0.2.0
nltk==3.8.1
safetensors==0.4.3
spacy==3.7.5